Update strings.xml
I added some translations of the new SAF and Library Categories stuffs in this update, and some more polishing of details.

I'm really sorry to trouble you, but the "Blacklist" and "Advanced" header in Settings -  Other was not found in String. So I added them in it. Could you please make them inside? Thanks a lot!
For example,
    <string name="pref_header_advanced">Advanced</string>
    <string name="pref_header_blacklist">Blacklist</string>

And also, the new "Not more than 5 items." was not found either, if it could be in String, I'll be really grateful!
Thanks!