Playback Speed and Pitch labels and StepSize changes and fixes.
PLAYBACK SPEED AND PITCH CHANGELOG:

Slider stepSize changed to: 0.01x
Because depending on the audio, stepping by 0.05x can cause cracking noises to the audio and make it the kind of robotic. Stepping by 0.01x the user has more control over the playback speed and can also eliminate the cracking noises.

Fixed the decimal places for the Playback Speed and Pitch. Depending on the range, it was displaying decimal values with more than 2 places.

Example:
Before: 0.9855566
With the fixes applied: 0.98

PS: I decided to use "DecimalFormat("#.##").format(value)" and "DecimalFormat("#.0#").format(value)" To avoid any issues with the user's native language decimal point formatting system, as this app is downloaded worldwide.

Playback Speed and Pitch labels, now display 1.0 insted of 1.00.

Instead of 0.25x up to 4.0x, now Playback Speed, has a range between: 0.75x up to 2.0x. Because these are the most used and necessary ones.